Move emlib submodules to modules as part of the library itself. Rewritten makefiles to create a crate for emlib, and move the code from src/lib.rs to example/main.rs.
