+++ Makefile
@@ -1,7 +1,7 @@
 DEVICE=EFM32GG990F1024
 TARGET=thumbv7m-none-eabi
 
-PROJ_FILE=lib
+PROJ_FILE=example/main
 PROJ_NAME=buttons_int
 
 -include Makefile.user
@@ -35,7 +35,7 @@ RUSTFLAGS += -L . -L $(LIB_DIR) --verbose
 FLASHFLAGS = --verify --reset
 
 
-%.elf: src/$(PROJ_FILE).rs $(LIB_DIR)/libcompiler-rt.a $(LIB_DIR)/libcore.rlib
+%.elf: $(PROJ_FILE).rs $(LIB_DIR)/libcompiler-rt.a $(LIB_DIR)/libcore.rlib $(LIB_DIR)/libemlib.rlib
 	$(RUSTC) $(RUSTFLAGS) $< -o $@
 
 %.hex: %.elf
